requirements.txt
streamlit
spotipy
Pillow
